+++ packages/web/src/App.tsx
@@ -3,6 +3,7 @@ import Layout from "./components/Layout";
 import HomePage from "./pages/HomePage";
 import PrivacyPage from "./pages/PrivacyPage";
 import TermsPage from "./pages/TermsPage";
+import LegalPage from "./pages/LegalPage";
 import SeoLandingPage from "./pages/SeoLandingPage";
 import NotFoundPage from "./pages/NotFoundPage";
 
@@ -13,6 +14,7 @@ export default function App() {
         <Route path="/" element={<HomePage />} />
         <Route path="/privacy" element={<PrivacyPage />} />
         <Route path="/terms" element={<TermsPage />} />
+        <Route path="/legal" element={<LegalPage />} />
         <Route path="/exif-remover" element={<SeoLandingPage slug="exif-remover" />} />
         <Route path="/gps-remover-photo" element={<SeoLandingPage slug="gps-remover-photo" />} />
         <Route path="/pdf-metadata-remove" element={<SeoLandingPage slug="pdf-metadata-remove" />} />